Fix for https://github.com/coralproject/pillar/issues/21. Consolidated code into a single go file.
